refactor: reorganize package structure and consolidate core client logic into new modular components
