+++ .gitignore
@@ -19,3 +19,4 @@ local.properties
 /.idea/
 /.kotlin/
 /group connect.txt
+.agents
