+++ app/src/main/java/com/sendspinlite/SendspinApplication.kt
@@ -1,5 +1,6 @@
 package com.sendspinlite
 
+import com.sendspinlite.diagnostics.CrashReportingManager
 import android.app.Application
 
 /**
