+++ app/src/main/java/com/sendspinlite/diagnostics/ReportingUtils.kt
@@ -1,4 +1,4 @@
-package com.sendspinlite
+package com.sendspinlite.diagnostics
 
 /** Returns the last [n] lines of [text] joined by newlines. */
 internal fun lastLines(
