+++ app/src/main/java/com/sendspinlite/service/SendspinService.kt
@@ -1,4 +1,4 @@
-package com.sendspinlite
+package com.sendspinlite.service
 
 import android.app.*
 import android.app.ActivityManager
@@ -23,6 +23,10 @@ import android.util.Log
 import androidx.core.app.NotificationCompat
 import kotlinx.coroutines.*
 import kotlinx.coroutines.flow.MutableStateFlow
+import com.sendspinlite.client.ClientEvent
+import com.sendspinlite.client.SendspinPcmClient
+import com.sendspinlite.ui.MainActivity
+import com.sendspinlite.ui.PlayerViewModel
 import kotlinx.coroutines.flow.StateFlow
 
 class SendspinService : Service() {
@@ -182,48 +186,6 @@ class SendspinService : Service() {
                 if (state.status.startsWith("failure:") && reconnectJob == null) {
                     Log.i(tag, "Connection failed, starting auto-reconnect")
                     startAutoReconnect(state.wsUrl, state.clientId, state.clientName)
-                }
-            }
-        }
-
-        // Monitor for server-commanded volume/mute changes and apply them to system
-        // This allows volume control to work even when MainActivity is not running (e.g., after boot)
-        scope.launch {
-            _uiState.collect { state ->
-                // Handle server-commanded volume changes
-                if (state.playerVolumeFromServer) {
-                    val audioManager = getSystemService(Context.AUDIO_SERVICE) as android.media.AudioManager
-                    val maxVolume = audioManager.getStreamMaxVolume(android.media.AudioManager.STREAM_MUSIC)
-                    val systemVolume = (state.playerVolume * maxVolume / 100)
-                    audioManager.setStreamVolume(android.media.AudioManager.STREAM_MUSIC, systemVolume, 0)
-                    Log.i(tag, "Applied server volume command: ${state.playerVolume}% (systemVolume=$systemVolume)")
-                    markServerVolumeSet()
-
-                    // Clear the flag
-                    _uiState.value = _uiState.value.copy(playerVolumeFromServer = false)
-                }
-
-                // Handle server-commanded mute changes
-                if (state.playerMutedFromServer) {
-                    if (state.playerMuted) {
-                        val audioManager = getSystemService(Context.AUDIO_SERVICE) as android.media.AudioManager
-                        audioManager.setStreamVolume(android.media.AudioManager.STREAM_MUSIC, 0, 0)
-                        Log.i(tag, "Applied server mute command: muted=${state.playerMuted}")
-                    }
-                    markServerVolumeSet()
-
-                    // Clear the flag
-                    _uiState.value = _uiState.value.copy(playerMutedFromServer = false)
-                }
-
-                // Handle server-commanded static delay changes: persist to SharedPreferences
-                if (state.staticDelayMsFromServer) {
-                    val prefs = getSharedPreferences("SendspinPlayerPrefs", Context.MODE_PRIVATE)
-                    prefs.edit().putLong("static_delay_ms", state.staticDelayMs).apply()
-                    Log.i(tag, "Persisted server-commanded static delay: ${state.staticDelayMs}ms")
-
-                    // Clear the flag
-                    _uiState.value = _uiState.value.copy(staticDelayMsFromServer = false)
                 }
             }
         }
@@ -494,41 +456,149 @@ class SendspinService : Service() {
             _uiState.value = _uiState.value.copy(staticDelayMs = savedStaticDelayMs)
         }
 
-        client =
+        val activeClient =
             SendspinPcmClient(
                 wsUrl = wsUrl,
                 clientId = clientId,
                 clientName = clientName,
                 context = this,
-                onUiUpdate = { patch ->
-                    val previousState = _uiState.value
-                    var newState = patch(previousState)
+            )
+        client = activeClient
+        activeClient.setStaticDelayMs(_uiState.value.staticDelayMs)
 
-                    // Mark that we had at least one successful websocket connection.
-                    if (newState.connected && newState.status == "ws_open") {
-                        hasEstablishedConnection = true
-                        dropCountedForCurrentOutage = false
-                    }
+        // Listen to client diagnostics Flow
+        scope.launch {
+            activeClient.diagnostics.collect { diag ->
+                val previousState = _uiState.value
+                var newState = previousState.copy(
+                    status = diag.status,
+                    connected = diag.connected,
+                    activeRoles = diag.activeRoles,
+                    playbackState = diag.playbackState,
+                    groupName = diag.groupName,
+                    streamDesc = diag.streamDesc,
+                    offsetUncertaintyUs = diag.offsetUncertaintyUs,
+                    driftPpm = diag.driftPpm,
+                    driftUncertaintyPpm = diag.driftUncertaintyPpm,
+                    driftSnr = diag.driftSnr,
+                    rttUs = diag.rttUs,
+                    networkQuality = diag.networkQuality,
+                    stability = diag.stability,
+                    connectionType = diag.connectionType,
+                    queuedChunks = diag.queuedChunks,
+                    bufferAheadMs = diag.bufferAheadMs,
+                    lateDrops = diag.lateDrops,
+                    audibleSyncCount = diag.audibleSyncCount,
+                    kalmanErrorCount = diag.kalmanErrorCount,
+                    groupVolume = diag.groupVolume,
+                    groupMuted = diag.groupMuted,
+                    supportedCommands = diag.supportedCommands,
+                    playbackSpeedMultiplier = diag.playbackSpeedMultiplier,
+                    smoothedLatencyMs = diag.smoothedLatencyMs,
+                    audioOutputStarted = diag.audioOutputStarted,
+                    playbackRecoveryStatus = diag.playbackRecoveryStatus,
+                    lastRecoveryEvent = diag.lastRecoveryEvent,
+                    clockReadyForPlayback = diag.clockReadyForPlayback,
+                    forceResyncActive = diag.forceResyncActive,
+                    inDiscontinuityRecovery = diag.inDiscontinuityRecovery,
+                    lateRestartLoops = diag.lateRestartLoops,
+                    effectiveBufferAheadMs = diag.effectiveBufferAheadMs,
+                    estimatedOffsetMs = diag.estimatedOffsetMs,
+                    playoutOffsetMs = diag.playoutOffsetMs,
+                    networkJitterMs = diag.networkJitterMs,
+                    clockUpdateCount = diag.clockUpdateCount,
+                    serverLatenessMs = diag.serverLatenessMs,
+                    lastAudioCutAgeMs = diag.lastAudioCutAgeMs,
+                    metadataTimestamp = diag.metadataTimestamp,
+                    trackTitle = diag.trackTitle,
+                    trackArtist = diag.trackArtist,
+                    albumTitle = diag.albumTitle,
+                    albumArtist = diag.albumArtist,
+                    trackYear = diag.trackYear,
+                    trackNumber = diag.trackNumber,
+                    artworkUrl = diag.artworkUrl,
+                    artworkBitmap = diag.artworkBitmap,
+                    trackProgress = diag.trackProgress,
+                    trackDuration = diag.trackDuration,
+                    playbackSpeed = diag.playbackSpeed,
+                    repeatMode = diag.repeatMode,
+                    shuffleEnabled = diag.shuffleEnabled,
+                    playerVolume = diag.playerVolume,
+                    playerVolumeFromServer = diag.playerVolumeFromServer,
+                    playerMuted = diag.playerMuted,
+                    playerMutedFromServer = diag.playerMutedFromServer,
+                    staticDelayMs = diag.staticDelayMs,
+                    staticDelayMsFromServer = diag.staticDelayMsFromServer
+                )
 
-                    // Count only unexpected connection losses, and only once per outage.
-                    val unexpectedDisconnect =
-                        hasEstablishedConnection &&
-                            previousState.connected &&
-                            !newState.connected &&
-                            (newState.status.startsWith("failure:") || newState.status.startsWith("closed:"))
+                // Mark that we had at least one successful websocket connection.
+                if (newState.connected && newState.status == "ws_open") {
+                    hasEstablishedConnection = true
+                    dropCountedForCurrentOutage = false
+                }
 
-                    if (unexpectedDisconnect && !dropCountedForCurrentOutage) {
-                        dropCountedForCurrentOutage = true
-                        newState = newState.copy(connectionDrops = previousState.connectionDrops + 1)
-                        Log.w(tag, "Unexpected connection drop detected. totalDrops=${newState.connectionDrops}, status=${newState.status}")
-                    }
+                // Count only unexpected connection losses, and only once per outage.
+                val unexpectedDisconnect =
+                    hasEstablishedConnection &&
+                        previousState.connected &&
+                        !newState.connected &&
+                        (newState.status.startsWith("failure:") || newState.status.startsWith("closed:"))
 
-                    _uiState.value = newState
-                    updateNotification()
-                },
-            ).also {
-                it.setStaticDelayMs(_uiState.value.staticDelayMs)
+                if (unexpectedDisconnect && !dropCountedForCurrentOutage) {
+                    dropCountedForCurrentOutage = true
+                    newState = newState.copy(connectionDrops = previousState.connectionDrops + 1)
+                    Log.w(tag, "Unexpected connection drop detected. totalDrops=${newState.connectionDrops}, status=${newState.status}")
+                }
+
+                _uiState.value = newState
+                updateNotification()
             }
+        }
+
+        // Listen to server events (Volume, Mute, Delay changes)
+        scope.launch {
+            activeClient.events.collect { event ->
+                when (event) {
+                    is ClientEvent.ServerVolumeChanged -> {
+                        val audioManager = getSystemService(Context.AUDIO_SERVICE) as android.media.AudioManager
+                        val maxVolume = audioManager.getStreamMaxVolume(android.media.AudioManager.STREAM_MUSIC)
+                        val systemVolume = (event.volume * maxVolume / 100)
+                        audioManager.setStreamVolume(android.media.AudioManager.STREAM_MUSIC, systemVolume, 0)
+                        Log.i(tag, "Applied server volume command: ${event.volume}% (systemVolume=$systemVolume)")
+                        markServerVolumeSet()
+
+                        // Sync volume to ViewModel via _uiState immediately
+                        _uiState.value = _uiState.value.copy(
+                            playerVolume = event.volume,
+                            playerVolumeFromServer = false
+                        )
+                    }
+                    is ClientEvent.ServerMutedChanged -> {
+                        val audioManager = getSystemService(Context.AUDIO_SERVICE) as android.media.AudioManager
+                        if (event.muted) {
+                            audioManager.setStreamVolume(android.media.AudioManager.STREAM_MUSIC, 0, 0)
+                            Log.i(tag, "Applied server mute command: muted=true")
+                        }
+                        markServerVolumeSet()
+
+                        _uiState.value = _uiState.value.copy(
+                            playerMuted = event.muted,
+                            playerMutedFromServer = false
+                        )
+                    }
+                    is ClientEvent.ServerStaticDelayChanged -> {
+                        val prefs = getSharedPreferences("SendspinPlayerPrefs", Context.MODE_PRIVATE)
+                        prefs.edit().putLong("static_delay_ms", event.delayMs).apply()
+                        Log.i(tag, "Persisted server-commanded static delay: ${event.delayMs}ms")
+
+                        _uiState.value = _uiState.value.copy(
+                            staticDelayMs = event.delayMs,
+                            staticDelayMsFromServer = false
+                        )
+                    }
+                }
+            }
+        }
 
         // Start health monitoring when connecting
         startHealthMonitoring()
@@ -656,12 +726,6 @@ class SendspinService : Service() {
         _uiState.value = _uiState.value.copy(staticDelayMs = clamped)
         client?.setStaticDelayMs(clamped)
     }
-
-    fun setEnableOpusCodec(enabled: Boolean) {
-        _uiState.value = _uiState.value.copy(enableOpusCodec = enabled)
-        client?.setEnableOpusCodec(enabled)
-    }
-
     // Player (local device) volume controls
     fun setPlayerVolume(volume: Int) {
         client?.setPlayerVolume(volume)
